requirements.txt
flask
flask-sqlalchemy
flask-wtf
email_validator
python-dotenv
flask-login